test: 로또 클래스 테스트 코드 추가
